+++ CHANGELOG.md
@@ -5,6 +5,7 @@ ADDED:
 - Ability to define options for `JsonApiClient`
 - `Document::error($number)` utility method to retrieve only one error
 - `Response::isSuccessful()` to determine if the response was successful
+- `Response::isSuccessful()` to determine if the response was successful and contains a document
 
 CHANGED: